requirements.txt
clip~=1.0
flax~=0.2.2
jax~=0.2.10
numpy~=1.19.5
tensorflow~=2.4.1
clu~=0.0.1a2
dataclasses~=0.6
scipy~=1.6.0
torchvision~=0.8.0
pillow~=7.2.0
torch~=1.7.0
timm~=0.4.5
tqdm~=4.49.0
opencv-python~=4.1.2.30
transformers~=4.5.1
eagerpy~=0.29.0
tensorboard~=2.4.1
matplotlib~=3.3.3
torchsummary~=1.5.1
pandas~=1.2.0